Version 2 progress report B
Extended Genre List:
The list of genres in the index function within app.py has been expanded to include a comprehensive list of genres provided	.

Enhanced CSS for Better Styling (Milestone 2)
The CSS (style.css) was updated to include more styling for a better user experience. This includes a cleaner layout, improved font, and better use of colors and spacing.

Added Bootstrap for Responsive Design
Bootstrap, a popular CSS framework, was added to the project to enhance the design and make the app responsive. This allows the app to look good on both desktop and mobile devices.

Improved HTML Structure
The HTML structure was improved by using Bootstrap classes. This helps in creating a more polished and responsive layout.